+++ src/main/java/com/example/std/service/DepartmentService.java
@@ -6,13 +6,13 @@ import com.example.std.model.Department;
 
 public interface DepartmentService {
 	
-	int savedepartment(Department department);
+	Long savedepartment(Department department);
 	
 	List<Department> getAllDepartments();
 	
-	Department getOneDepartment(Integer id);
+	Department getOneDepartment(Long id);
 	
-	void deleteDepartment(Integer id);
+	void deleteDepartment(Long id);
 
 	List<Department> getAllDepartmentsinPage(int number);
